revise compile-mgard-zchecker.sh

+++ mgard-patches/compile-mgard-zchecker.sh
@@ -1,8 +1,11 @@
 #!/bin/bash
-cp Makefile_LIBP_MGARD ../libpressio/test/Makefile
-cp mgarddouble_CompDecomp.c ../libpressio/test
-cp mgardfloat_CompDecomp.c ../libpressio/test
 
-cd ../libpressio/test
-make
+mgardpath=`abspath ../compressor-install`
+if [ ! -f "$mgardpath" ]; then
+	cp Makefile_LIBP_MGARD ../libpressio/test/Makefile
+	cp mgarddouble_CompDecomp.c ../libpressio/test
+	cp mgardfloat_CompDecomp.c ../libpressio/test
 
+	cd ../libpressio/test
+	make
+fi
